Added noteEdit feature

+++ NoteApp/Views/ListView.swift
@@ -6,16 +6,26 @@ import SwiftUI
 struct ListView : View {
     @EnvironmentObject var listViewModel : ListViewModel
     var body: some View {
-        List{
-            ForEach(listViewModel.items,){ item in
-                ListRowView(item: item).onTapGesture {
-                    // Open a NotePage 
+        ZStack{
+            if listViewModel.items.isEmpty{
+                EmptyItemView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            }else{
+                List{
+                    ForEach(listViewModel.items,){ item in
+                        NavigationLink(
+                            destination:
+                                AddView(currentItem: item).environmentObject(listViewModel))
+                        {
+                            ListRowView(item: item)
+                        }
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
                 }
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
-        .listStyle(PlainListStyle())
         .navigationTitle("Todo list 📋")
         .navigationBarItems(
             leading: EditButton(),
@@ -32,4 +42,12 @@ struct ListView : View {
     NavigationView {
         ListView()
     }.environmentObject(ListViewModel())
+}
+
+
+#Preview {
+    NavigationView {
+        ListView()
+    }.environmentObject(ListViewModel())
+        .preferredColorScheme(.dark)
 }
